+++ pesto/__init__.py
@@ -1,4 +1,3 @@
 from .core import load_model, predict, predict_from_files
 
-
-__version__ = '1.0.0'
+__version__ = "2.0.1"